+++ app/src/main/java/com/dede/android_eggs/views/main/compose/EasterEggItems.kt
@@ -22,6 +22,7 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ExperimentalLayoutApi
 import androidx.compose.foundation.layout.FlowRow
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
@@ -121,7 +122,12 @@ fun EasterEggHighestItem(
                 egg.provideSnapshotProvider()
             }
 
-            SnapshotView(snapshot)
+            SnapshotView(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .aspectRatio(1.8f),
+                snapshot = snapshot
+            )
 
             if (isSupportShortcut) {
                 IconButton(