+++ app/src/main/java/com/dede/android_eggs/views/settings/SettingsScreen.kt
@@ -55,6 +55,7 @@ import com.dede.android_eggs.views.settings.compose.prefs.LanguagePref
 import com.dede.android_eggs.views.settings.compose.prefs.LanguagePrefUtil
 import com.dede.android_eggs.views.settings.compose.prefs.RetainInRecentsPref
 import com.dede.android_eggs.views.settings.compose.prefs.RocketLauncherPref
+import com.dede.android_eggs.views.settings.compose.prefs.SnapshotPref
 import com.dede.android_eggs.views.settings.compose.prefs.ThemePref
 import com.dede.android_eggs.views.settings.compose.prefs.TimelinePref
 import kotlinx.coroutines.launch
@@ -138,6 +139,8 @@ fun SettingsScreen(drawerState: DrawerState = rememberDrawerState(DrawerValue.Cl
             }
 
             SettingDivider()
+
+            SnapshotPref()
 
             TimelinePref()