+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### v4.3.2
 
-- Add Android L Preview Easter Egg
+- Add Android L Preview Easter Egg [#714](https://github.com/hushenghao/AndroidEasterEggs/issues/714)
+- Add Platlogo snapshot preview
 - Optimization Cat Editor options panel
 - Fix Timeline list scroll gesture conflicts [#654](https://github.com/hushenghao/AndroidEasterEggs/issues/654)
 - Upgrade the minimum supported version to Android 6.0